+++ docs/examples/promise/90_typegen/src/interfaces/treasuryRewards/types.ts
@@ -1,0 +1,7 @@
+// Auto-generated via `yarn polkadot-types-from-defs`, do not edit
+/* eslint-disable @typescript-eslint/no-empty-interface */
+
+import { u128 } from '@polkadot/types/primitive';
+
+/** @name Balance2 */
+export interface Balance2 extends u128 {}